+++ R/add_q.R
@@ -11,6 +11,7 @@
 #' @inheritParams add_global_p.tbl_regression
 #' @author Esther Drill, Daniel D. Sjoberg
 #' @family tbl_summary tools
+#' @family tbl_svysummary tools
 #' @family tbl_regression tools
 #' @family tbl_uvregression tools
 #' @export
@@ -84,7 +85,8 @@ add_q <- function(x, method = "fdr", pvalue_fun = NULL, quiet = NULL) {
   # update table_header --------------------------------------------------------
   # footnote text
   footnote_text <-
-    add_q_method_lookup[add_q_method_lookup$method == method, ]$method_label
+    add_q_method_lookup[add_q_method_lookup$method == method, ]$method_label %>%
+    translate_text()
 
   x$table_header <-
     tibble(column = names(x$table_body)) %>%
@@ -99,7 +101,7 @@ add_q <- function(x, method = "fdr", pvalue_fun = NULL, quiet = NULL) {
     )
 
   # adding  column header
-  x <- modify_header_internal(x, q.value = "**q-value**")
+  x <- modify_header_internal(x, q.value = paste0("**", translate_text("q-value"), "**"))
 
   # return final object --------------------------------------------------------
   # adding call