+++ R/as_flextable.R
@@ -26,11 +26,11 @@
 #' Pro tip: Use the [flextable::width()] function for exacting control over
 #' column width after calling [as_flextable()].
 #' @inheritParams as_gt
+#' @inheritParams as_tibble.gtsummary
 #' @param strip_md_bold When TRUE, all double asterisk (markdown language for
 #' bold weight) in column labels and spanning headers are removed.
 #' Default is TRUE
 #' @param ... Not used
-#' @name as_flextable
 #' @export
 #' @return A {flextable} object
 #' @family gtsummary output types
@@ -41,16 +41,12 @@
 #'   tbl_summary(by = trt) %>%
 #'   add_p() %>%
 #'   as_flextable()
-as_flextable <- function(x, ...) {
-  UseMethod("as_flextable")
-}
-
-#' @rdname as_flextable
-#' @export
 as_flextable.gtsummary <- function(x, include = everything(), return_calls = FALSE,
-                                   strip_md_bold = TRUE, ...) {
-  # must have flextable package installed to use this function -----------------
-  assert_package("flextable", "as_flextable")
+                                   strip_md_bold = TRUE, group_header = NULL, ...) {
+  # setting defaults -----------------------------------------------------------
+  group_header <-
+    group_header %||%
+    get_theme_element("pkgwide-str:group_header", default = "**Group**")
 
   # stripping markdown asterisk ------------------------------------------------
   if (strip_md_bold == TRUE) {
@@ -60,10 +56,11 @@ as_flextable.gtsummary <- function(x, include = everything(), return_calls = FAL
         vars(.data$label, .data$spanning_header),
         ~str_replace_all(., pattern = fixed("**"), replacement = fixed(""))
       )
+    group_header <- str_replace_all(group_header, pattern = fixed("**"), replacement = fixed(""))
   }
 
   # creating list of flextable calls -------------------------------------------
-  flextable_calls <- table_header_to_flextable_calls(x = x)
+  flextable_calls <- table_header_to_flextable_calls(x = x, group_header = group_header)
 
   # adding user-specified calls ------------------------------------------------
   insert_expr_after <- get_theme_element("as_flextable.gtsummary-lst:addl_cmds")
@@ -96,12 +93,27 @@ as_flextable.gtsummary <- function(x, include = everything(), return_calls = FAL
 }
 
 # creating flextable calls from table_header -----------------------------------
-table_header_to_flextable_calls <- function(x, ...) {
-  table_header <-
-    x$table_header %>%
-    group_by(.data$hide) %>%
-    mutate(id = ifelse(.data$hide == FALSE, dplyr::row_number(), NA)) %>%
-    ungroup()
+table_header_to_flextable_calls <- function(x, group_header, ...) {
+
+  # if there is a grouping variable, add table_header info for it
+  if (dplyr::group_vars(x$table_body) %>% length() > 0) {
+    table_header <-
+      tibble::tibble(column = "groupname_col",
+                     label = group_header,
+                     hide = FALSE,
+                     align = "left") %>%
+      bind_rows(x$table_header) %>%
+      group_by(.data$hide) %>%
+      mutate(id = ifelse(.data$hide == FALSE, dplyr::row_number(), NA)) %>%
+      ungroup()
+  }
+  else {
+    table_header <-
+      x$table_header %>%
+      group_by(.data$hide) %>%
+      mutate(id = ifelse(.data$hide == FALSE, dplyr::row_number(), NA)) %>%
+      ungroup()
+  }
 
   # tibble ---------------------------------------------------------------------
   # flextable doesn't use the markdown language `__` or `**`