+++ R/tbl_merge.R
@@ -1,7 +1,7 @@
 #' Merge two or more gtsummary objects
 #'
 #' Merges two or more `tbl_regression`, `tbl_uvregression`, `tbl_stack`,
-#' or `tbl_summary` objects and adds appropriate spanning headers.
+#' `tbl_summary`, or `tbl_svysummary` objects and adds appropriate spanning headers.
 #'
 #' @param tbls List of gtsummary objects to merge
 #' @param tab_spanner Character vector specifying the spanning headers.
@@ -11,6 +11,7 @@
 #' @family tbl_regression tools
 #' @family tbl_uvregression tools
 #' @family tbl_summary tools
+#' @family tbl_svysummary tools
 #' @seealso [tbl_stack]
 #' @author Daniel D. Sjoberg
 #' @export
@@ -88,8 +89,9 @@ tbl_merge <- function(tbls, tab_spanner = NULL) {
 
   # merging tables -------------------------------------------------------------
   # nesting data by variable (one line per variable), and renaming columns with number suffix
-  nested_table <- tbls %>%
-    imap(function(x, y) {
+  nested_table <- map2(
+    tbls, seq_along(tbls),
+    function(x, y) {
       # creating a column that is the variable label
       group_by(x$table_body, .data$variable) %>%
         mutate(
@@ -106,7 +108,7 @@ tbl_merge <- function(tbls, tab_spanner = NULL) {
   # nesting results within variable
   nested_table <- map(
     nested_table,
-    ~ nest(.x, data = -one_of(c("variable", "var_label")))
+    ~ nest(.x, data = -any_of(c("variable", "var_label")))
   )
 
   # merging formatted objects together
@@ -148,8 +150,8 @@ tbl_merge <- function(tbls, tab_spanner = NULL) {
 
   # stacking all table_header dfs together and renaming ------------------------
   table_header <-
-    imap_dfr(
-      tbls,
+    purrr::map2_dfr(
+      tbls, seq_along(tbls),
       ~ pluck(.x, "table_header") %>%
         # tidying the code in these columns (giving it space to breathe),
         # that is can be properly pasred in the next step