+++ tests/testthat/test-add_overall.R
@@ -24,3 +24,35 @@ test_that("no errors/warnings with missing data", {
 test_that("no errors/warnings with missing data in by variable", {
   expect_error(trial %>% tbl_summary(by = response) %>% add_overall(), NA)
 })
+
+test_that("add_overall-works with ordered factors", {
+  expect_error(
+    trial %>%
+      select(response, trt) %>%
+      dplyr::mutate_at(vars(response, trt),
+                       ~factor(., ordered = TRUE)) %>%
+      tbl_summary(by = trt) %>%
+      add_overall(),
+    NA
+  )
+})
+
+test_that("no errors/warnings with standard use for tbl_svysummary", {
+  t <- trial %>%
+    survey::svydesign(data = ., ids = ~ 1, weights = ~ 1) %>%
+    tbl_svysummary(by = trt)
+
+  expect_error(t %>% add_overall(), NA)
+  expect_warning(t %>% add_overall(), NA)
+
+  expect_error(t %>% add_overall(last = TRUE), NA)
+  expect_warning(t %>% add_overall(last = TRUE), NA)
+
+  t <- Titanic %>%
+    as.data.frame() %>%
+    survey::svydesign(data = ., ids = ~ 1, weights = ~ Freq) %>%
+    tbl_svysummary(by = Survived)
+
+  expect_error(t %>% add_overall(), NA)
+  expect_warning(t %>% add_overall(), NA)
+})